@@ -61,11 +64,9 @@
     "de-indent": "^1.0.2",
     "es6-promise": "^4.1.0",
     "escodegen": "^1.8.1",
-    "eslint": "^4.13.1",
-    "eslint-loader": "^1.7.1",
+    "eslint": "^5.7.0",
     "eslint-plugin-flowtype": "^2.34.0",
     "eslint-plugin-jasmine": "^2.8.4",
-    "eslint-plugin-vue-libs": "^2.0.1",
     "file-loader": "^1.1.5",
     "flow-bin": "^0.61.0",
     "hash-sum": "^1.0.2",
@@ -93,21 +94,20 @@
     "nightwatch-helpers": "^1.2.0",
     "phantomjs-prebuilt": "^2.1.14",
     "resolve": "^1.3.3",
-    "rollup": "^0.54.1",
+    "rollup": "^0.66.6",
     "rollup-plugin-alias": "^1.3.1",
-    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-babel": "^4.0.1",
     "rollup-plugin-buble": "^0.19.2",
-    "rollup-plugin-commonjs": "^8.0.0",
+    "rollup-plugin-commonjs": "^9.2.0",
     "rollup-plugin-flow-no-whitespace": "^1.0.0",
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-replace": "^2.0.0",
-    "rollup-watch": "^4.0.0",
     "selenium-server": "^2.53.1",
     "serialize-javascript": "^1.3.0",
     "shelljs": "^0.8.1",
-    "typescript": "^2.7.1",
-    "uglify-js": "^3.0.15",
-    "webpack": "^3.11.0",
+    "terser": "^3.10.2",
+    "typescript": "^3.1.3",
+    "webpack": "^4.22.0",
     "weex-js-runtime": "^0.23.6",
     "weex-styler": "^0.3.0",
     "yorkie": "^1.0.1"
