@@ -1,6 +1,6 @@
 {
   "name": "@cafee/staff",
-  "version": "1.0.12",
+  "version": "1.1.3",
   "description": "",
   "main": "dist/main.js",
   "scripts": {
@@ -25,7 +25,7 @@
     "importz": "^1.0.10",
     "lodash": "^4.17.11",
     "npm-process": "^1.1.0",
-    "promise-super": "^1.0.3",
+    "promise-super": "^1.1.5",
     "ressive": "^2.1.2",
     "vue": "^2.5.17",
     "vue-i18n": "^8.4.0",
@@ -33,13 +33,14 @@
     "vuex": "^3.0.1"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-0": "^6.24.1",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
+    "@babel/runtime": "^7.2.0",
+    "babel-loader": "^8.0.4",
     "cross-env": "^5.2.0",
     "cross-env-argv": "^1.0.5",
-    "webpack": "^3.11.0",
+    "webpack": "^4.27.1",
     "webpack-root": "^1.1.3"
   }
 }
